requirements.txt
altgraph==0.17.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
customtkinter==5.2.2
darkdetect==0.8.0
docutils==0.20.1
filetype==1.2.0
flask==3.0.3
Flask-SocketIO==5.5.1
h11==0.16.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
Kivy==2.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==25.0
pillow==10.4.0
pycparser==2.22
pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
simple-websocket==1.1.0
urllib3==2.2.3
werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
